requirements.txt
PyQt6>=6.6
matplotlib>=3.8

# YAML config parsing
pyyaml>=6.0

# Web API
fastapi>=0.116
uvicorn>=0.30
httpx>=0.27     # needed for FastAPI TestClient


# PDF generation
reportlab>=4.0
